Refactor deployment form; add deployment parameters section component
