+++ src/components/DeploymentForm.vue
@@ -54,29 +54,7 @@
 
       <p-divider />
 
-      <p-content>
-        <h3 class="deployment-form__section-header">
-          Parameters
-        </h3>
-
-        <p-button-group v-model="parametersInput" :options="parametersInputOptions" size="sm" />
-
-        <template v-if="hasParameters">
-          <template v-if="parametersInput === 'form'">
-            <DeploymentParameters v-model="parameters" :deployment="deployment" />
-          </template>
-
-          <template v-else>
-            <p-label :state="jsonParametersState" :message="jsonParametersErrorMessage">
-              <p-code-input v-model="jsonParameters" lang="json" :min-lines="3" show-line-numbers />
-            </p-label>
-          </template>
-        </template>
-
-        <template v-else>
-          <em>This deployment's flow has no parameters</em>
-        </template>
-      </p-content>
+      <DeploymentParametersSection v-model="parameters" :deployment="deployment" />
     </p-content>
 
     <p-divider />
@@ -96,7 +74,6 @@
       </p-label>
     </p-content>
 
-
     <template #footer>
       <p-button inset @click="cancel">
         Cancel
@@ -109,28 +86,18 @@
 </template>
 
 <script lang="ts" setup>
-  import { useValidation } from '@prefecthq/vue-compositions'
-  import { merge } from 'lodash'
   import { useField } from 'vee-validate'
-  import { computed, ref } from 'vue'
-  import { ScheduleFieldset, WorkPoolCombobox, DeploymentParameters, WorkPoolQueueCombobox, JsonInput } from '@/components'
+  import { computed } from 'vue'
+  import { DeploymentParametersSection, ScheduleFieldset, WorkPoolCombobox, WorkPoolQueueCombobox, JsonInput } from '@/components'
   import { useForm } from '@/compositions/useForm'
   import { localization } from '@/localization'
   import { Deployment, DeploymentUpdate, DeploymentEdit, Schedule } from '@/models'
-  import { getSchemaDefaultValues, mapper } from '@/services'
   import { SchemaValues } from '@/types/schemas'
-  import { stringify, isJson, fieldRules, stringifyUnknownJson, parseUnknownJson, isRecord } from '@/utilities'
+  import { stringify, isJson, fieldRules } from '@/utilities'
 
   const props = defineProps<{
     deployment: Deployment,
   }>()
-
-  const parametersInputOptions = [{ value: 'form', label: 'Form' }, { value: 'json', label: 'JSON' }]
-  const parametersInput = ref<'form' | 'json'>('form')
-
-  const hasParameters = computed(() => {
-    return Object.keys(props.deployment.parameterOpenApiSchema.properties ?? {}).length > 0
-  })
 
   const name = computed(() => props.deployment.name)
 
@@ -162,37 +129,12 @@
   const { value: tags } = useField<string[] | null>('tags')
   const { value: infrastructureOverrides, meta: overrideState, errorMessage: overrideErrorMessage } = useField<string>('infrastructureOverrides', rules.infrastructureOverrides)
 
-  const parameterOpenApiSchema = computed(() => {
-    const { rawSchema } = props.deployment
-
-    if (rawSchema && 'required' in rawSchema) {
-      rawSchema.required = []
-    }
-
-    return mapper.map('SchemaResponse', rawSchema ?? {}, 'Schema')
-  })
-  const jsonParameters = ref(stringifyUnknownJson(merge(getSchemaDefaultValues(parameterOpenApiSchema.value), props.deployment.rawParameters)))
-  const { error: jsonParametersErrorMessage, state: jsonParametersState, validate: validateJsonParameters } = useValidation(jsonParameters, localization.info.parameters, rules.jsonParameters)
-
   const emit = defineEmits<{
     (event: 'submit', value: DeploymentUpdate): void,
     (event: 'cancel'): void,
   }>()
 
   const submit = handleSubmit(async (values): Promise<void> => {
-    if (parametersInput.value == 'json') {
-      const valid = await validateJsonParameters()
-
-      if (!valid) {
-        return
-      }
-
-      const parsed = parseUnknownJson(jsonParameters.value)
-      if (isRecord(parsed)) {
-        values.parameters = parsed
-      }
-    }
-
     const deploymentUpdate: DeploymentUpdate = {
       ...values,
       infrastructureOverrides: JSON.parse(infrastructureOverrides.value),
