+++ src/components/index.ts
@@ -72,6 +72,7 @@ export { default as DeploymentList } from './DeploymentList.vue'
 export { default as DeploymentListItem } from './DeploymentListItem.vue'
 export { default as DeploymentMenu } from './DeploymentMenu.vue'
 export { default as DeploymentParameters } from './DeploymentParameters.vue'
+export { default as DeploymentParametersSection } from './DeploymentParametersSection.vue'
 export { default as DeploymentQuickRunOverflowMenuItem } from './DeploymentQuickRunOverflowMenuItem.vue'
 export { default as DeploymentsCount } from './DeploymentsCount.vue'
 export { default as DeploymentsDeleteButton } from './DeploymentsDeleteButton.vue'
@@ -91,6 +92,7 @@ export { default as FlowListItem } from './FlowListItem.vue'
 export { default as FlowMenu } from './FlowMenu.vue'
 export { default as FlowRouterLink } from './FlowRouterLink.vue'
 export { default as FlowRunArtifacts } from './FlowRunArtifacts.vue'
+export { default as FlowRunBreadCrumbs } from './FlowRunBreadCrumbs.vue'
 export { default as FlowRunCancelButton } from './FlowRunCancelButton.vue'
 export { default as FlowRunCancelModal } from './FlowRunCancelModal.vue'
 export { default as FlowRunCreateForm } from './FlowRunCreateForm.vue'
@@ -232,7 +234,6 @@ export { default as TimezoneSelect } from './TimezoneSelect.vue'
 export { default as ToastFlowRunCreate } from './ToastFlowRunCreate.vue'
 export { default as VariableCreateModal } from './VariableCreateModal.vue'
 export { default as VariableEditModal } from './VariableEditModal.vue'
-export { default as FlowRunBreadCrumbs } from './FlowRunBreadCrumbs.vue'
 export { default as VariableMenu } from './VariableMenu.vue'
 export { default as VariablesDeleteButton } from './VariablesDeleteButton.vue'
 export { default as VariablesTable } from './VariablesTable.vue'
